Separate out opcodes & stack behaviour
